cloud flares work loaders
this is for cloud flares work loaders as a load balancer. 
In this updated code, the handleRequest function will keep trying to forward the request to the next server in the servers array until it gets a successful response, or until it has tried all of the servers. If all of the servers are down or if there is an error during the request or response process, the function will return a response with a 503 status code and a "Service Unavailable" message.